Updated openstack/openstack Project: openstack/ironic  337956f3bc4bafcca7ee9342eff1c0162fbbff9e
Include graphviz in install prerequisites

Graphviz is needed for generating the state machine diagram
(eg, via 'tox -egenstates'). It is added to the list of packages
to install.

The instructions for installing on openSUSE and SLE are a bit
more involved, so I've added a URL that describes how to do that.

Change-Id: I16380bc88fcecc6911f21b4554b04a9d72b2d66b