Updated openstack/openstack Project: openstack/governance  b70b43bb9f696add6079aecfa9c7fc4c293d9311
Add "bifrost" to Ironic project

Bifrost is a set of ansible playbooks and roles that demonstrate how to
configure and use Ironic to provision hardware, without requiring any
other OpenStack services.

This is useful as a CI tool and as a minimal deployment tool for small
environments where shared resource management is not needed.

Change-Id: I2552803a1da44312a0e8879629036a23d89e2e2a