Updated openstack/openstack Project: openstack/tooz  353f5ba82313b33c8818ebb4a8897875ba2232a7
Use a serialization/deserialization specific exception

Instead of raising the top level tooz error on loads/dumps
failures, instead raise a specific one that is only used
when serialization/deserialization has some sort of issue.

This also changes the exception raised in the memcache driver
to also use this new exception type (when a unknown serialization
format is encountered).

Change-Id: Ic4302b1bf3dba45b70af85a081bebbe0512c6830