Updated openstack/openstack Project: openstack/nova  306f39ef3d32baecdf7725900acaba1aeadfde19
console: clean tokens do not happen for all kind of consoles

The current condition to verify whether or not consoles are
enable and so flush tokens when destroying instance is not
kept sync with every kind of consoles that can be enable.

This commit create one entry point to clean tokens + update the
missing serial console config.

Already covered by tests.

Closes-Bug: #1451801
Change-Id: Ie7cb03a9431549df75d7fbc82cf19e3f8f410a0f