Updated openstack/openstack Project: openstack-infra/devstack-gate  9562e2855ab2d44b5ffb6cc03235fb4b0ebf1511
Check for mount before creating swap

The existing umount appears to be just for safety, as it does not
actually unmount anything.  It fails with errexit, but rather than
ignore this, it seems better to check if there is a mount and bail
out.

If somehow the ephemeral storage was mounted, it implies somebody set
it up somewhere else and wouldn't be expecting this to go ahead and
repartition it on them.

Change-Id: Id7d345130b751973d997940d7b62a4c253d265ed