Updated openstack/openstack Project: openstack-infra/project-config  63dbb3d406021388064c12e4bad4a863fa192216
Look for yum instead of rpm when caching devstack

On CentOS 6.x the rpm executable is in /bin instead of /usr/bin, but
cache_devstack.py ends up running yum if it wants to download RPM
packages anyway. Look for yum instead of rpm as an indication of
which packages to install.

Change-Id: Iad79c0fcf66d2bd457195f007009c76f6e6aa2d2