Updated openstack/openstack Project: openstack-infra/system-config  e24e1b5b99060aeb5983fe4aac45266e6c11bbfd
Revert "Revert zuul status in gerrit"

This reverts commit ff6d73b989c57db79bb53e82f814d651945c3ac6.

This brings back zuul inline status which is further optimized in
follow on patches. We do this in phases because it's a lot easier to
review that way.

It also introduces zuul_inline global control variable, so we can land
this turned off, and enable in the debugger to verify it's doing the
correct thing.

Change-Id: If951a4a05a7dfef7a6d0c8e15ed263ff0523165e