Updated openstack/openstack Project: openstack-infra/devstack-gate  8eddc7326c284a87c7c79f56fa774a0bb7781f0c
Add host counter to ansible inventory

Add host counter to the inventory. This is usedful to keep track of
known node offsets for log captures and overlay network setup.

Make use of the host counter to copy subnode logs to a known locations
to prove that this works as expected.

Change-Id: I2d2368d1546332ed1edc21eb0275df5bd8618825