Updated openstack/openstack Project: openstack-dev/devstack  872cdf62ce862e4971b9bfa8e20cd7bd8a9bb679
Set local_ip only when TENANT_TUNNELS are enabled

In an installation with VLAN tenant networks, devstack should
not configure the local_ip (which is applicable only when
tenant_tunnels are used). This is causing failures in Neutron
for an IPv6 only setup. This patch addresses this issue, but
configuring the local_ip only when TENANT_TUNNELS are enabled.

Related-Bug: #1447693
Change-Id: I0e2a2d8b6ce0ad87f6c0d318ac522dbab50d44ee