Updated openstack/openstack Project: openstack/horizon  936e190c18ee19ba4926d3ee59854ea14067b6b5
Fixed dashboard angular translation issue

We now have javascript files in openstack_dashboard as well. However,
the run_test script does not account for this and only run translation on
the horizon folder. This patch address this problem. Also removed the
invalid gettext('') since that causes the msgid to be blank.

Steps to test:
1. ./run_tests.sh --makemessages
2. ./run_tests.sh --pseudo de
3. ./run_tests.sh --compilemessages
4. set LAUNCH_INSTANCE_NG_ENABLED = True in local_settings.py
5. localhost:8001/settings/ and change to deutsch
6. localhost:port/project/instances/ and click on launch instance

Co-Authored-By: Doug Fish <drfish@us.ibm.com>
Change-Id: If1cd1937d40d359a4a4ecd75406f62e8f8802082
Closes-bug: #1444150