Updated openstack/openstack Project: openstack/neutron-vpnaas  5400931e658602fdde3dc2ce2e3a8512d22829ae
IPv6 support for OpenSwan, Libreswan and Strongswan

VPNaaS is not working with IPv6 addresses.
This patch will address the below issues and make VPNaas
work with IPv6 addresses

1)ipsec doesn't need nat rule for ipv6. So not adding this for ipv6.
This issue is applicable for openSwan, libreswan and strongswan.

The below two issues are seen only on openswan and libreswan.
No changes are required to strongswan template files to support ipv6.
2)Add 'connaddrfamily=ipv6' to ipsec.conf for whack to parse all
addresses in the config as ipv6 addresses.

3)openswan fails to process ipv6 address from %defaultroute, so
explicitly pass the gateway address as leftnexthop address.

Change-Id: I74832be4476bd36cdcd4234b30d0daee24ef61ce
Closes-bug: #1435136
Closes-bug: #1436263
Closes-bug: #1436890