Updated openstack/openstack Project: openstack/nova  977798c5026f39dfa69bc042a4b3e2f0ff18f37b
Cleanup conductor unused methods

There are some conductor api methods left unused that can be removed
together with two very old manager methods that were deprecated and
scheduled for removal with the rpc api version bump to 2.0.

This change removes the following unused api client methods:

 - instance_get_all_by_host()
 - instance_get_all_by_host_and_node()
 - migration_get_in_progress_by_host_and_node()
 - aggregate_metadata_get_by_host()
 - block_device_mapping_create()
 - block_device_mapping_update()
 - block_device_mapping_update_or_create()
 - block_device_mapping_get_all_by_instance()
 - compute_node_update()
 - compute_node_delete()
 - security_groups_trigger_handler()

Furthermore, it removes these previously deprecated manager methods:

 - ping()
 - instance_get_active_by_window()

Manager methods that are deprecated and scheduled for removal on the
next rpc version bump are annotated with a todo comment if not there
already.

Change-Id: Ia825cee94e6ba9431a85c1b6f64cb26e7c83c95f