Updated openstack/openstack Project: openstack/nova  3aefd933de6fac66337d2b782248fffdbb153dc2
Merge multinic functional tests between v2 and v2.1

Currently v2 and v2.1 have separate functional tests and their
corresponding sample files. As v2 and v2.1 are supposed to be identical,
there is overhead to maintain two set of functional tests and sample files.
We can have one set of tests which can run for both v2 and v2.1.

This commit merges multinic functional tests.

In V2.1 os-access-ip is separate plugin where in v2 it was mixed in server
plugin itself. So os-access-ip extension is enabled to run
merged functional tests for both v2 and v2.1.

Change-Id: I92c2d128ccb5728de77991544a58c1b6c6ca47fc