Updated openstack/openstack Project: openstack/nova  23948b0a7bf54888b3904d1ad8476f118a8e6f20
Merge server usage tests between v2 and v2.1

Currently v2 and v2.1 have separate functional tests and their
corresponding sample files. As v2 and v2.1 are supposed to be identical,
there is overhead to maintain two set of functional tests and sample files.
We can have one set of tests which can run for both v2 and v2.1.

This commit merges servers usage functional tests.

In V2.1 os-access-ip is separate plugin where in v2 it was mixed in server
plugin itself. So os-access-ip extension is enabled to run
merged functional tests for both v2 and v2.1.

Change-Id: I46bfeaed8f47ea7674f14d6f3d3b301dcc35d419