Updated openstack/openstack Project: openstack/nova  1c292f333740dc5b8784d5ed19c898bbc47b6ce5
servicegroup: remove get_all method never used as public

Retrieve all services is exposed in to few circumstances to
justify to maintain a method get_all in every drivers of
servicegroup instead we have to continue to use
Service.get_all_by_topics() (which is used by default)
Also this method is never used as a public and only the
zk driver needs it.
Finally it makes pain for some drivers like Memcached to
maintain a routine like that.

This commit removes the method get_all from the public API
and rename the method to be private for the Zookeeper's
driver.

Change-Id: Idc0dfbbe1887e11166acb7d989dd5466751761af