Updated openstack/openstack Project: openstack/neutron  cf8a76743361bb96d714339efb6b5c514d9eeeb4
changes log level to debug for help calls

Calling help typically causes the program being called to exit non-zero.
This causes the command to be logged as an error even though it should not be.
By setting 'log_fail_as_error=False' we log to debug.  This helps clean
up logs.

Change-Id: I13f9488b9bc524bb85047e9b2dcf4e8a76bd6c11
Closes-Bug: 1452425