Updated openstack/openstack Project: openstack/heat  798f205a79b1944e31f9f0aeec8f9eede51115f6
Update template paths for environment-mapped TemplateResources

When updating a template resource, we weren't taking into account the
possibility that the new environment could be mapping the resource to a
different template path. This change ensures we recalculate the template
name from the new environment before reloading the template during an
update.

Change-Id: I5f84f6fd60925d3dba0b8e8dec867657c8c5c9ef
Closes-Bug: #1452534