Updated openstack/openstack Project: openstack-infra/project-config  47532ee5c35baf47dc37895fac32fc86f6666b7e
Install yum-plugin-downloadonly on nodes

The --downloadonly option to yum is provided by the
yum-plugin-downloadonly package. This is merely a virtual package
satisfied by yum itself in newer releases, but an optional package
in older ones such as CentOS/RHEL 6.x. Install it just to be sure it
will work, since we use this to pre-cache RPMs on nodepool images.

Change-Id: I9e2e1605f3721c410180aa46a81b7b731d08503a