Updated openstack/openstack Project: openstack-infra/elastic-recheck  d8f406955f6eccea3b3dc24c82d4720b53084dcd
Add context "libvirt" back into query for bug 1451506

In https://review.openstack.org/#/c/180684/ the query was updated to
catch more formats of the same error, but it removed the context
of "libvirt" from the query. This change makes the query also match
on "libvirt" in the error message.

Related-Bug: #1451506

Change-Id: Iae26a0c14e46243e3c3f6eeb7f470cc4d5dedef3