Updated openstack/openstack Project: openstack-infra/nodepool  98f1ce2634be18cc3d6f5433c8efc857b4f93e93
Allow for no ZMQ publishers

You could not start nodepoold without zmq publishers because publisher
reconfiguration properly detected that it did not need to reconfigure
any publishers but then failed when it tried to use the non existant
self.config.zmq_publishers to set the new config's publishers.

Fix this by wrapping with a check for self.confg, this is how the
gearman client deals with this issue.

Change-Id: I50a24c2786ddadf3f37e8093244551c756470768