Updated openstack/openstack Project: openstack/horizon  c1c3f66875daf595f71ce3498eb4a979f8743ccd
Fixing Jasmine tests in dashboard

We currently have two jasmine tests, one in horizon and one in dashboard.
We currently use two different templates and they are out of sync with each
other. Furthermore, we should extend the sources list, not override it,
because dashboard modules depends on code in horizon.

This patch fixes these issues.
Closes-bug: #1452446
Change-Id: I3fcbb1ea30b5e83a16f94fb31959270bd393593a