Updated openstack/openstack Project: openstack/neutron  bde93dfef9f170cc20ca09770b88c7e1c43c3ff2
VMWare NSXv: Metadata for distributed router

Metadata support for NSXv distributed routers is provided via DHCP Edge
appliances. In order to avoid conflicts between distributed routers and
DHCP Edges which map different networks with same CIDRs, we create a 1:1
mapping between an distributed router and a DHCP Edge.

This patch contains the data model for the above, while the implementation
is in vmware-nsx repository.

Change-Id: I324403f7d5df4861193840e05bedf7a473aea655