Updated openstack/openstack Project: openstack-infra/nodepool  e25f65a17798ca80d8a2e173baf20f041351cdb9
Allow for Jenkinsless targets

Nodepool intends to be runnable without a Jenkins to set up as a target.
This broke because there are two places in the code that don't check for
jenkins config in the target before attempting to get a Jenkins Manager
for that target. Fix this by wrapping these cases with a check for
target.jenkins_url (the same check used elsewhere in nodepool).

You do still need to configure a target with a name though, just omit
the jenkins configuration.

Change-Id: I22949d85c04bcf282adfa5095820908a4130bd7d