Updated openstack/openstack Project: openstack-infra/zuul  d7650853007205b72fae6e83dd8f4fbeb8d0a9e4
Fix race in test_crd_check_duplicate more

The previous fix was insufficient because multiple merge jobs
could be run at once.  Using the new orderedRelease method however
should be sufficient, and it ensures only one job runs and finishes
at a time.

The previous code was more appropriate for a test of a gate
pipeline and was probably copied from it.

Change-Id: I0534c7ce6dd473cb71ba37c8c0fedef8e5ca72a5