Updated openstack/openstack Project: openstack/requirements  40b86ac9cb707ff97f5b140d4f5fe5f1dcbbb2ed
Bump pbr versions.

We're about to release pbr 1.0. This needs to be permitted :). Also
anything < 0.10 was totally broken with PEP-440 in pip, so we have
to bump the lower end. Since we're bumping it, bring everything in
line with the current release - which we gate on everywere anyway
because setup_requires is unversioned.

Change-Id: Ia7205b4599ae648cdb4935111458d92f2065b47e