Updated openstack/openstack Project: openstack/taskflow  e6a04193e73bd44450eca9859b035cf91b4fcf4b
Avoid trying to copy tasks results when cloning/copying

It is not always safe to copy task results so instad of
trying to copy.copy them just retain the existing results
in the clone (and avoid copying any users results if
they exist).

This appeared to be causing sqlalchemy weak references to
get garbage collected prematurely in cinder.

Fixes bug 1452978

Change-Id: I75e4ad0e9eb50284ff09179565b6d146a8da3d92