Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  6c6370bbff20ff40631c12430268e9dd4d8bef2f
Query all jobs once

Instead of doing an individual api call each time,
to check if a job exists, do an initial jenkins get_jobs()
call on init, and use that list as a cache, to speed
up job queries.

If a job is not found in cache, we still will be querying
jenkins directly as a fallback.

Change-Id: Ie3eff875f4d832a42c9e38478f62055bc135a7e0