Updated openstack/openstack Project: openstack-infra/system-config  e27649729ee7b5f6d8f989cb6f4a2cc46b4e5972
Avoid race on reload_gerrit_header

When updating Javascript included by Gerrit's site header, the way
to get it reprocessed is to update the timestamp of the site header
file. However anecdotal tests indicate that if the file's timestamp
is updated too soon after the Javascript in question, this does not
help. Introduce a 10-second delay between any Javascript replacement
and updating the timestamp on the site header file to avoid this
race.

While we're here, correct an obvious typographical error in the
executable path for the calling environment.

Change-Id: I8ebd6d8f6dee227cd4f2c3820b40972bcda4ebce