Updated openstack/openstack Project: openstack-infra/project-config  fb9472006bf7a820db8888328a6487c42281ec26
Only import files if there are new translations

Do not import files if only content is deleted but no new translation
added. If there's at least one file with added translations, files with
removed content are also imported.

This should reduce the number of imports, since a project should see an
import only if new strings are in transifex.

Change-Id: Id521a2209bf9cad3b90acf81447db027beed7b64