Updated openstack/openstack Project: openstack/barbican  01417a89ca2cb0c99b81e4a07989d49fd4038e8b
Migrate to oslo_context

This CR introduces oslo_context as a dependency, in order to be more
aligned with the rest of OpenStack. Although, due to the fact that we
had some extra fields in the context, the approach that I took was to
inherit from oslo_context's RequestContext and add the extra fields that
we needed.

I also removed the unused fields (which are related to the ownership and
the service catalog), which lead to also deleting the context middleware
tests; Since it was only one test and the functionality that it was
testing is no longer needed.

Change-Id: Id439a32f0d21fe7d8adafec71417a3e9d4870ec5