Updated openstack/openstack Project: openstack/nova  c7b9ecef78e40994295fe7b907db9dae72d31c89
Log the actual instance.info_cache when empty in floating ip associate

This attempts to add some enlightenment to the case where we try to
associate a floating ip for an instance and hit the case where there
is no info_cache. We need to know if it's actually None, or empty.

Related-Bug: #1249065

Change-Id: Ic0051d1d520f800d845b7006b5b95be5d8edf7c9