Updated openstack/openstack Project: openstack/cinder  571fa5f9b62655a25165ad1aeb5002783d1772b3
Correct directories check for N327 hacking check

For some reason the N327 hacking heck was looking for
cinder.unit/tests and cinder.functional/tests in the filename
that is sent in for the hacking check.  This should be
cinder/tests/unit and cinder/tests/functional.

This concern was originally raised in this review:
https://review.openstack.org/175604

Change-Id: Ic3041d9e16b3707aaf42ff941206d24ec66ead2b