Updated openstack/openstack Project: openstack/cinder-specs  ca54b61129c64d14695b36a5e28f2d7047c9ace2
Fix the volume replication spec

The spec that was originally written for volume replication
was written assuming that Cinder would provide the bulk of the
work to implement volume replication.

After the spec was approved it was decided that a different
approach would be taken to implement volume replication.  It
was decided that Cinder would just provide a framework for
backend drivers to implement volume replication as appropriate
for their storage backend with an extra spec set in the volume type
to indicate that replication should be used for a volume.  The code that was
merged in Juno implemented volume replication using this approach.

Unfortunately, the spec was not updated after these changes were
made to the design which has caused confusion for developers wishing
to implement volume replication.

This commit brings the spec up to date with the merged functionality.

Change-Id: I56311d0471123635197ebfdeb1547007e75bd3f8