Updated openstack/openstack Project: openstack/nova  f9e664f1b6521dd7d5c02cd803e376e8abdf9c30
unify libvirt driver checks for qemu

There are multiple features that are only available with qemu.
It makes no sense to have a separate constant for each feature
when they're all looking for the same thing.

There is currently no list of libvirt hypervisor drivers, so
we start such a list.  This is not intended to be exhaustive,
but rather only for those drivers that need special handling.

Change-Id: I3b9546debb3c5bdca13926005a7664c701c6fb9a