Updated openstack/openstack Project: openstack/barbican  0ee250bd07e44cd980f7030e7af166edb88b26b8
Expose root cause plugin exceptions

Stevedore was being used to both discover configured plugins and to
create instances of these plugins. If the plugin's __init__() function
raised an exception however, stevedore logged the exception but
otherwise kept processing plugins in the list. The logger also is
disabled so no stack trace is seen in the log file, hence the init
error is swallowed. It wasn't until the plugin was searched for later
in the flow that an error was appearent to the client, but the original
root cause of the issue was not found. This CR creates these plugins
outside of stevedore so that issues are immediately raised for
attention, as these represent configuration issues that must be fixed
for proper operation.

Change-Id: I2fd533302ac5e0364aee6e24fd03fcf962c7a3ef