Updated openstack/openstack Project: openstack-infra/puppet-meetbot  1190b252c26b24706986d1b721fddb247eae12b9
Produce html logs for our IRC logs

There is a tool, irclog2html, that knows how to take directories of irc
logs and turn them in to formatted html with per-line links. Since we do
that for build logs, and the logs associated with meetings, it seems
only sane that we do it for channel logs.

The tool groks mtime, so it will skip previously produced logs, and will
overwrite current streaming logs each time it's run. We'll need to run
the command one time by hand first, because the first run might take a
bit to eat through our history.

Change-Id: Ieed2c80d945ee35b29887d672cd9ebbd9ece2188