Updated openstack/openstack Project: openstack/nova  5bd31c958331b39206567bbab59f2ba8128395b8
Update kilo version alias

Kilo actually sends 4.0 by default and can support 3.x and 4.x messages.
Deployers upgrading from Juno->Kilo can set the compute upgrade level to
'juno' until all nodes are upgraded to kilo and then unset the pin. When
moving to Liberty, setting the level to 'kilo' will use 4.x messages as
Liberty does not support 3.x anymore.

Note that this changes all the use of the kilo alias in test_rpcapi to
just use 3.40 (what it was before). This makes the tests pass (because
it's the same as we're expecting) for the backport. These tests are
removed in the next patch anyway.

UpgradeImpact

Change-Id: Ibbe83879ee1ccfab582119b1e2e588f0471020e2