Updated openstack/openstack Project: openstack/heat  2b7d0c2314c0383c271b81ebca89a039eb1b97d0
Do not specify version for contrib plugins

The pbr 0.11.0 release has shown that having multiple trees in a repo
each with their on version that is different to the git history is
really not going to work.

This change means that the version will be derived from git history,
which will still be useful for operators managing what version of a
plugin they have installed.

Running sudo python setup.py install (as documented) is known to work
with the latest pbr.

Change-Id: I716a5374bb283d11e26459b5866a0437ab6fe0dd
Closes-Bug: #1450733