Updated openstack/openstack Project: openstack/python-ironicclient  3f4671a0b7eb5c300291c2e92082d16b96e6316c
Cleanup session creation

When using the session object we want to pass through unknown kwargs to
the adapter so that keystoneclient can add new parameters as required.

This requires fixing the old options so that we can restrict what is
passed and showing a warning when there are values that are being
ignored.

Change-Id: Ia795dfc8e16c60a08df1636816daefe2ebbd0b03