Updated openstack/openstack Project: openstack/oslo.log  c694f9aa26866f6103a1ca9f017f88f5649b8a4e
Log deprecation message when catching deprecated exceptions

When an exception is being deprecated from a library, it is usually
not called anywhere in the user code -- instead it is only caught, so
we never get a chance to see the deprecation message. This patch uses
the fact that an except statement has to do a subclass check on the
exceptions it is trying to catch, and puts the logging code in the
subclass check hook. Unfortunately, that hook has to be defined on the
metaclass, which makes the code a little hard to understand.

Change-Id: Id3cb649343c8d32564e189aecb9aeebe572249cc