Updated openstack/openstack Project: openstack/api-site  5a11f8343cd0253c18ca81f704a822bc3be62d6a
Publish Libcloud version of FirstApp

The Libcloud version of FirstApp is ready to move from draft stage
so it can be widely disseminated to satiate the burning need for this
content that is currently not being addressed. The guide was very well
received during it's packed-room summit sesion.

So far it has received significantly more editing and testing than
other sprint-created documents when they were published. This is thanks
to the work of: Andreas Jaegar, Karen Bradshaw, Diane Fleming, Shilla Saebi,
James Dempsey's users, Jon Proulx, Bruno Lago, Chuck Burkett, Dana Bauer,
and the original authors (at least 2 each section).

Any areas lacking in the guide are well understood and captured in the
bug list:
https://bugs.launchpad.net/openstack-manuals/+bugs?field.tag=firstapp
and there are no major blockers that would prevent listing this guide
on developer.openstack.org.

This patch links the libcloud section of developer.openstack.org to the
guide.

Change-Id: If85bc020cd54e3ca06e5d320024dec9428764641